Use BigDecimal for ticket prices in DTOs
Replaced `Double` with `BigDecimal` in ticket-related request/response DTOs to improve monetary precision and avoid floating-point rounding issues. Also removed an unused import and added missing `@NotNull` validation on `start` and `end` in `UpdateEventRequestDTO` to enforce required event dates.

+++ src/main/java/com/danny/ticket/domain/dtos/ListEventTicketTypeResponseDTO.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.lang.classfile.constantpool.DoubleEntry;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -14,7 +14,7 @@ public class ListEventTicketTypeResponseDTO {
 
     private UUID id;
     private String name;
-    private Double price;
+    private BigDecimal price;
     private String description;
     private Integer totalAvailable;
